refactor(p2): extract Common message UI helpers into sibling page (stage 1 of 4)
Moves 4 shared private helpers from ChatMessage.kt into new sibling
ChatMessageCommon.kt (same package): TraceChatComposable + MessageSelectionContainer +
rememberClickCitationHandler + MessageAnnotations.

Byte-equal: bodies and signatures unchanged.
Visibility:
- TraceChatComposable: private (only same-sibling caller MessageSelectionContainer)
- MessageSelectionContainer: private -> internal (Stage 2 VAT + Stage 4 MessagePartsBlock)
- rememberClickCitationHandler: private -> internal (Stage 2 VAT + Stage 4 MessagePartsBlock)
- MessageAnnotations: private -> internal (Stage 3 CMVF + Stage 4 MessagePartsBlock)

Coordinator strip: 1634 -> 1504 (-130 with import cleanup).
Imports trimmed: 16 orphans (Trace, SelectionContainer, TextButton, stringResource,
LinkAnnotation, buildAnnotatedString, withLink, jsonArray, jsonObject, jsonPrimitive,
BuildConfig, Favicon, extendColors, JsonInstant, openUrl, urlDecode).
Pre-existing dead preserved: getValue, setValue (used implicitly via `by` in coord),
fastForEach (already 0 body refs in baseline 3809e414).

Baseline 3/467 unit tests preserved.
